+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-*
+* Ability to duplicate an existing print job, reusing its original payload, from both the API and the Admin UI
+* Ability to download a print job's original payload from the Admin UI
 
 ### Changed